ai-engine/requirements.txt
fastapi
uvicorn
jinja2
python-multipart
openai
pdfplumber
pymupdf
pytesseract
pillow
python-docx